+++ TabletopTweaks-Base/Bugfixes/Classes/Slayer.cs
@@ -1,9 +1,13 @@
 ﻿using HarmonyLib;
+using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.JsonSystem;
 using Kingmaker.EntitySystem.Stats;
 using Kingmaker.UnitLogic.FactLogic;
 using Kingmaker.UnitLogic.Mechanics;
+using Kingmaker.Utility;
+using System.Linq;
+using TabletopTweaks.Core.NewComponents.Prerequisites;
 using TabletopTweaks.Core.Utilities;
 using static TabletopTweaks.Base.Main;
 
@@ -18,9 +22,41 @@ namespace TabletopTweaks.Base.Bugfixes.Classes {
                 Initialized = true;
                 TTTContext.Logger.LogHeader("Patching Slayer");
 
-                PatchBaseClass();
+                PatchAlternateCapstone();
+                PatchBase();
             }
-            static void PatchBaseClass() {
+            static void PatchAlternateCapstone() {
+                if (Main.TTTContext.Fixes.AlternateCapstones.IsDisabled("Skald")) { return; }
+
+                var MasterSlayerFeature = BlueprintTools.GetBlueprintReference<BlueprintFeatureBaseReference>("a26c0279a423fc94cabeea898f4d9f8a");
+                var SlayerAlternateCapstone = NewContent.AlternateCapstones.Slayer.SlayerAlternateCapstone.ToReference<BlueprintFeatureBaseReference>();
+
+                MasterSlayerFeature.Get().TemporaryContext(bp => {
+                    bp.AddComponent<PrerequisiteInPlayerParty>(c => {
+                        c.CheckInProgression = true;
+                        c.HideInUI = true;
+                        c.Not = true;
+                    });
+                    bp.HideNotAvailibleInUI = true;
+                    TTTContext.Logger.LogPatch(bp);
+                });
+                ClassTools.Classes.SlayerClass.TemporaryContext(bp => {
+                    bp.Progression.UIGroups
+                        .Where(group => group.m_Features.Any(f => f.deserializedGuid == MasterSlayerFeature.deserializedGuid))
+                        .ForEach(group => group.m_Features.Add(SlayerAlternateCapstone));
+                    bp.Progression.LevelEntries
+                        .Where(entry => entry.Level == 20)
+                        .ForEach(entry => entry.m_Features.Add(SlayerAlternateCapstone));
+                    bp.Archetypes.ForEach(a => {
+                        a.RemoveFeatures
+                            .Where(remove => remove.Level == 20)
+                            .Where(remove => remove.m_Features.Any(f => f.deserializedGuid == MasterSlayerFeature.deserializedGuid))
+                            .ForEach(remove => remove.m_Features.Add(SlayerAlternateCapstone));
+                    });
+                    TTTContext.Logger.LogPatch("Enabled Alternate Capstones", bp);
+                });
+            }
+            static void PatchBase() {
                 PatchSlayerTrapfinding();
 
                 void PatchSlayerTrapfinding() {