+++ TabletopTweaks-Base/NewContent/AlternateCapstones/Warpriest.cs
@@ -1,5 +1,8 @@
-﻿using Kingmaker.Blueprints.Classes;
+﻿using Kingmaker.Blueprints;
+using Kingmaker.Blueprints.Classes;
 using Kingmaker.Blueprints.Classes.Selection;
+using Kingmaker.UnitLogic.FactLogic;
+using System.Collections.Generic;
 using TabletopTweaks.Core.NewComponents.Prerequisites;
 using TabletopTweaks.Core.Utilities;
 using static TabletopTweaks.Base.Main;
@@ -8,13 +11,36 @@ namespace TabletopTweaks.Base.NewContent.AlternateCapstones {
     internal class Warpriest {
         public static BlueprintFeatureSelection WarpriestAlternateCapstone = null;
         public static void AddAlternateCapstones() {
+            var SecondBlessingSelection = BlueprintTools.GetBlueprint<BlueprintFeatureSelection>("b7ce4a67287cda746a59b31c042305cf");
+            var BlessingResource = BlueprintTools.GetBlueprintReference<BlueprintAbilityResourceReference>("d128a6332e4ea7c4a9862b9fdb358cca");
             var WarpriestAspectOfWar = BlueprintTools.GetBlueprint<BlueprintFeature>("65cc7abc21826a344aa156e2a40dcecc");
 
-            var HammerOfGod = Helpers.CreateBlueprint<BlueprintFeature>(TTTContext, "HammerOfGod", bp => {
+            var HammerOfGodBlessingSelection = SecondBlessingSelection.CreateCopy(TTTContext, "HammerOfGodBlessingSelection", bp => {
+                bp.Group = FeatureGroup.None;
+            });
+            var HammerOfGod = Helpers.CreateBlueprint<BlueprintProgression>(TTTContext, "HammerOfGod", bp => {
                 bp.SetName(TTTContext, "Hammer of God");
                 bp.SetDescription(TTTContext, "At 20th level, the warpriest has become one of his deity’s favorite weapons—the first tool that comes to hand when destruction is called for.\n" +
                     "The warpriest gains two additional blessings from the list offered by his deity. He can also call upon his blessings two more times each day.");
+                bp.Ranks = 1;
                 bp.IsClassFeature = true;
+                bp.GiveFeaturesForPreviousLevels = true;
+                bp.ReapplyOnLevelUp = true;
+                bp.HideNotAvailibleInUI = true;
+                bp.m_Classes = new BlueprintProgression.ClassWithLevel[0];
+                bp.m_Archetypes = new BlueprintProgression.ArchetypeWithLevel[0];
+                bp.m_ExclusiveProgression = new BlueprintCharacterClassReference();
+                bp.m_FeaturesRankIncrease = new List<BlueprintFeatureReference>();
+                bp.LevelEntries = new LevelEntry[] {
+                    Helpers.CreateLevelEntry(20,
+                        HammerOfGodBlessingSelection,
+                        HammerOfGodBlessingSelection
+                    )
+                };
+                bp.AddComponent<IncreaseResourceAmount>(c => {
+                    c.m_Resource = BlessingResource;
+                    c.Value = 2;
+                });
             });
             WarpriestAlternateCapstone = Helpers.CreateBlueprint<BlueprintFeatureSelection>(TTTContext, "WarpriestAlternateCapstone", bp => {
                 bp.SetName(TTTContext, "Capstone");